pyproject.toml
[project]
name = "Project-Name"
version = "0.1.0"
description = ""
authors = [
    {name = "Viresh Ambardar",email = "ambardarviresh18@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "numpy (>=2.4.1,<3.0.0)",
    "fastapi[standard] (>=0.135.2,<0.136.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "chromadb (>=1.0.20,<2.0.0)",
    "psutil (>=7.2.2,<8.0.0)",
    "langchain (>=1.2.15,<2.0.0)",
    "langchain-groq (>=1.1.2,<2.0.0)",
    "ipython (>=9.12.0,<10.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"



[tool.poetry]
package-mode = false


packages = [
    { include = "app" }
]

[tool.poetry.scripts]
start = "app.main:main"

